Déplacement du HTMLConverter dans le package util Refactor du HTMLConverter Ajout de tests pour HTMLConverter Mise à jour du pom pour cibler java11
